requirements.txt
streamlit==1.31.0
altair==4.2.2
pandas
biopython
fpdf
plotly
scipy
numpy
matplotlib
lxml
langchain-ollama